# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent API",
    "description": "API with listing and markets agents",
    "version": "0.1.0"
  },
  "paths": {
...
    "/guarantors": {
      "get": {
        "tags": [
          "guarantors"
        ],
        "summary": "Read Saved Guarantors",
        "operationId": "read_saved_guarantors_guarantors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGuarantorListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "guarantors"
        ],
        "summary": "Create Saved Guarantor Route",
        "operationId": "create_saved_guarantor_route_guarantors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedGuarantorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGuarantorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantors/{guarantor_id}": {
      "patch": {
        "tags": [
          "guarantors"
        ],
        "summary": "Patch Saved Guarantor Route",
        "operationId": "patch_saved_guarantor_route_guarantors__guarantor_id__patch",
        "parameters": [
          {
            "name": "guarantor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Guarantor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedGuarantorPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedGuarantorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "guarantors"
        ],
        "summary": "Delete Saved Guarantor Route",
        "operationId": "delete_saved_guarantor_route_guarantors__guarantor_id__delete",
        "parameters": [
          {
            "name": "guarantor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Guarantor Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantors/requests": {
      "get": {
        "tags": [
          "guarantors"
        ],
        "summary": "Read Guarantor Requests",
        "operationId": "read_guarantor_requests_guarantors_requests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuarantorRequestListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "guarantors"
        ],
        "summary": "Create Guarantor Request Route",
        "operationId": "create_guarantor_request_route_guarantors_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuarantorRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuarantorRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantors/requests/{request_id}": {
      "patch": {
        "tags": [
          "guarantors"
        ],
        "summary": "Patch Guarantor Request Route",
        "operationId": "patch_guarantor_request_route_guarantors_requests__request_id__patch",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuarantorRequestPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuarantorRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "guarantors"
        ],
        "summary": "Delete Guarantor Request Route",
        "operationId": "delete_guarantor_request_route_guarantors_requests__request_id__delete",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantors/requests/{request_id}/invite": {
      "post": {
        "tags": [
          "guarantors"
        ],
        "summary": "Invite Guarantor Request Route",
        "operationId": "invite_guarantor_request_route_guarantors_requests__request_id__invite_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuarantorInviteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantors/requests/{request_id}/decision": {
      "post": {
        "tags": [
          "guarantors"
        ],
        "summary": "Decision Guarantor Request Route",
        "operationId": "decision_guarantor_request_route_guarantors_requests__request_id__decision_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuarantorDecisionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuarantorRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantor-invite/{token}": {
      "get": {
        "tags": [
          "guarantors-public"
        ],
        "summary": "Read Guarantor Invite",
        "operationId": "read_guarantor_invite_guarantor_invite__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuarantorInviteContextOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantor-invite/{token}/open": {
      "post": {
        "tags": [
          "guarantors-public"
        ],
        "summary": "Open Guarantor Invite",
        "operationId": "open_guarantor_invite_guarantor_invite__token__open_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantor-invite/{token}/consent": {
      "post": {
        "tags": [
          "guarantors-public"
        ],
        "summary": "Consent Guarantor Invite",
        "operationId": "consent_guarantor_invite_guarantor_invite__token__consent_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuarantorInviteConsentIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantor-invite/{token}/sign": {
      "post": {
        "tags": [
          "guarantors-public"
        ],
        "summary": "Sign Guarantor Invite",
        "operationId": "sign_guarantor_invite_guarantor_invite__token__sign_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuarantorInviteSignIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantor-invite/{token}/documents": {
      "post": {
        "tags": [
          "guarantors-public"
        ],
        "summary": "Upload Guarantor Invite Documents",
        "operationId": "upload_guarantor_invite_documents_guarantor_invite__token__documents_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuarantorInviteDocumentIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/guarantor-invite/{token}/decline": {
      "post": {
        "tags": [
          "guarantors-public"
        ],
        "summary": "Decline Guarantor Invite",
        "operationId": "decline_guarantor_invite_guarantor_invite__token__decline_post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuarantorInviteDeclineIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GuarantorDecisionPatch": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "One of: verified, failed, declined, revoked"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "GuarantorDecisionPatch"
      },
      "GuarantorInviteConsentIn": {
        "properties": {
          "consent_text_version": {
            "type": "string",
            "title": "Consent Text Version"
          }
        },
        "type": "object",
        "required": [
          "consent_text_version"
        ],
        "title": "GuarantorInviteConsentIn"
      },
      "GuarantorInviteContextOut": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "guarantor_name": {
            "type": "string",
            "title": "Guarantor Name"
          },
          "guarantor_email": {
            "type": "string",
            "title": "Guarantor Email"
          },
          "lease": {
            "$ref": "#/components/schemas/LeasePayload"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "invite_expires_at": {
            "type": "string",
            "title": "Invite Expires At"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "guarantor_name",
          "guarantor_email",
          "lease",
          "status",
          "invite_expires_at"
        ],
        "title": "GuarantorInviteContextOut"
      },
      "GuarantorInviteDeclineIn": {
        "properties": {
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "title": "GuarantorInviteDeclineIn"
      },
      "GuarantorInviteDocumentIn": {
        "properties": {
          "document_type": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Document Type"
          },
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Content Type"
          },
          "byte_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Byte Size"
          },
          "storage_key": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Storage Key"
          },
          "metadata_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Json"
          }
        },
        "type": "object",
        "required": [
          "document_type",
          "filename",
          "content_type",
          "byte_size",
          "storage_key"
        ],
        "title": "GuarantorInviteDocumentIn"
      },
      "GuarantorInviteOut": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "invite_expires_at": {
            "type": "string",
            "title": "Invite Expires At"
          },
          "invite_url": {
            "type": "string",
            "title": "Invite Url"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "status",
          "invite_expires_at",
          "invite_url"
        ],
        "title": "GuarantorInviteOut"
      },
      "GuarantorInviteSignIn": {
        "properties": {
          "signer_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Signer Name"
          },
          "signer_email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Signer Email"
          },
          "signature_text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Signature Text"
          },
          "consent_text_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Consent Text Version"
          }
        },
        "type": "object",
        "required": [
          "signer_name",
          "signer_email",
          "signature_text",
          "consent_text_version"
        ],
        "title": "GuarantorInviteSignIn"
      },
      "GuarantorRequestCreate": {
        "properties": {
          "guarantor_id": {
            "type": "string",
            "title": "Guarantor Id"
          },
          "lease": {
            "$ref": "#/components/schemas/LeasePayload"
          }
        },
        "type": "object",
        "required": [
          "guarantor_id",
          "lease"
        ],
        "title": "GuarantorRequestCreate"
      },
      "GuarantorRequestListResponse": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/GuarantorRequestOut"
            },
            "type": "array",
            "title": "Requests"
          }
        },
        "type": "object",
        "required": [
          "requests"
        ],
        "title": "GuarantorRequestListResponse"
      },
      "GuarantorRequestOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "guarantor_id": {
            "type": "string",
            "title": "Guarantor Id"
          },
          "guarantor_snapshot": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Guarantor Snapshot"
          },
          "lease": {
            "$ref": "#/components/schemas/LeasePayload"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "sent_at": {
            "type": "string",
            "title": "Sent At"
          },
          "viewed_at": {
            "type": "string",
            "title": "Viewed At"
          },
          "signed_at": {
            "type": "string",
            "title": "Signed At"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "signing_events": {
            "items": {
              "$ref": "#/components/schemas/SigningEventOut"
            },
            "type": "array",
            "title": "Signing Events"
          }
        },
        "type": "object",
        "required": [
          "id",
          "guarantor_id",
          "guarantor_snapshot",
          "lease",
          "status",
          "verification_status",
          "created_at",
          "sent_at",
          "viewed_at",
          "signed_at",
          "expires_at",
          "signing_events"
        ],
        "title": "GuarantorRequestOut"
      },
      "GuarantorRequestPatch": {
        "properties": {
          "lease": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeasePayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "GuarantorRequestPatch"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SavedGuarantorCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "default": ""
          },
          "relationship": {
            "type": "string",
            "title": "Relationship",
            "default": "other"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "title": "SavedGuarantorCreate"
      },
      "SavedGuarantorListResponse": {
        "properties": {
          "saved_guarantors": {
            "items": {
              "$ref": "#/components/schemas/SavedGuarantorOut"
            },
            "type": "array",
            "title": "Saved Guarantors"
          }
        },
        "type": "object",
        "required": [
          "saved_guarantors"
        ],
        "title": "SavedGuarantorListResponse"
      },
      "SavedGuarantorOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "relationship": {
            "type": "string",
            "title": "Relationship"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "relationship",
          "created_at"
        ],
        "title": "SavedGuarantorOut"
      },
      "SavedGuarantorPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "relationship": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship"
          }
        },
        "type": "object",
        "title": "SavedGuarantorPatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}